lesson 6: move binding logic to viewholder
